fix: create useIsGlxOpen hook to separate logic from component
Use the hook where needed to get the values. TODO: Add test
